Updated navigation to nest the stack navigtor in the tab navigator - now can link to anything in the stack without having to put links in the BottomTabs
